requirements.txt
Django~=4.2.2
psycopg2-binary
scapy
dj-database-url
python-dotenv
reportlab
xhtml2pdf
pandas
numpy
matplotlib
django-chartjs
beautifulsoup4
lxml
celery
django-celery-beat
django-celery-results
django-kombu
pika
django-celery-results